create new classes and transfer code
