+++ LINQ/Helpers/WorkWithData.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace LINQ.Helpers
+{
+    public static class WorkWithData
+    {
+
+    }
+}
